requirements.txt
black==25.1.0
click==8.2.1
colorama==0.4.6
flake8==7.3.0
greenlet==3.2.3
iniconfig==2.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
